+++ CHANGELOG.md
@@ -93,6 +93,14 @@
     - `getSiteRef` retrieves a list of included areas for the site
     - `setLandingRef` sets the included areas for the page
     - `setSiteRef` sets the included areas for the site
+- Added service `Services\CRM\Documentgenerator\Numerator` with support methods,
+  see [crm.documentgenerator.numerator.* methods](https://github.com/bitrix24/b24phpsdk/issues/215):
+    - `add` adds a new numerator, with batch calls support
+    - `list` gets the list of numerators, with batch calls support
+    - `update` updates an existing numbering with new values, with batch calls support
+    - `delete` deletes a numerator, with batch calls support
+    - `get` gets information about the numerator by its identifier
+    - `count` count numerators
 - Added service `Services\Paysystem\Handler\Service\Handler` with support methods,
   see [pay_system.handler.* methods](https://github.com/bitrix24/b24phpsdk/issues/260):
     - `add` adds a payment system handler